+++ lib/rspec/rails/example/helper_example_group.rb
@@ -49,7 +49,10 @@ module RSpec::Rails
       # Returns an instance of ActionView::Base with the helper being specified
       # mixed in, along with any of the built-in rails helpers.
       def helper
-        _view.tap {|v| v.assign(view_assigns)}
+        _view.tap do |v|
+          v.extend(ApplicationHelper) if defined?(ApplicationHelper)
+          v.assign(view_assigns)
+        end
       end
 
     private